[Twinkle-#50][Swing] Use new method in panels

+++ twinkle-app/twinkle.main/src/main/java/com/github/donkeyrit/twinkle/panels/nestedpanels/PageNavigatorPanel.java
@@ -20,7 +20,7 @@ public class PageNavigatorPanel extends JPanel {
 		NavigationPanelEventsListener navigationPanelEventsListener, 
 		int chosenPage, 
 		int itemsPerPage, 
-		int maxItems
+		long maxItems
 	) {
 		this.navigationPanelEventsListener = navigationPanelEventsListener;
         this.chosenPage = chosenPage;
